Revert "moving ValidatorExtensions into root namespace and deleting ExtensionMethods folder from project."
This reverts commit 2fcf3b3d321c4895412455f9239819f055a1e6dd.
