+++ Validator/IsUrl.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using Validator.ExtensionMethods;
 using static System.Int32;
 
 namespace Validator
