+++ Validator/Validator.csproj
@@ -41,6 +41,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="ExtensionMethods\ValidatorExtensions.cs" />
     <Compile Include="IpVersion.cs" />
     <Compile Include="IsBase64.cs" />
     <Compile Include="IsbnVersion.cs" />
@@ -54,7 +55,6 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="UuidVersion.cs" />
     <Compile Include="Validator.cs" />
-    <Compile Include="ValidatorExtensions.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
